Move filter functions to new permanent home
